fix: quote the svd decomposition in a precision the build does not decide
The guide's singular-value block and the example script that mirrors it read
`torch.linalg.svdvals` of a float32 matrix and quoted it rounded to six
decimals. The exact value is 5.464985704219..., so the sixth decimal's
rounding boundary is 5.4649855 and the value sits 2.04e-7 from it -- inside
float32's own error at that magnitude. macOS/Accelerate computes 5.46498537
and prints 5.464985; Linux/OpenBLAS rounds the other way and prints 5.464986,
which is what CI reported on all three cpu-core interpreters.

Read the reference in double precision instead: the value is then 2.04e-7 from
the boundary against a double-precision error of about 6e-16, and the quoted
digit is the correctly rounded 5.464986 rather than a build's artifact. The
guide's two prose mentions of the same quantity become 5.4649857, a 7-decimal
truncation, so no two places give a different sixth decimal for one number.

The checker keeps its exact comparison, deliberately: a quoted value that
depends on the installed BLAS is a defect worth reporting, and a relative
tolerance would have hidden this one. Its docstring said the comparison's
premise was that every block seeds its sampler -- the wrong premise, since the
value that broke is a classical LAPACK result no seed governs. It now carries
the portability half, the measurement across macOS/Accelerate and
Linux/OpenBLAS on both architectures, and the limit that it cannot tell a
build disagreement from a real change.

+++ examples/algorithms/svd.py
@@ -57,7 +57,11 @@ def main() -> None:
     parser.add_argument("--seed", type=int, default=11)
     args = parser.parse_args()
 
-    exact = [float(value) for value in torch.linalg.svdvals(MATRIX)]
+    # Read in double precision, so the digits printed below are the matrix's rather than
+    # the installed BLAS's: in float32 the largest lands 2e-7 from the sixth decimal's
+    # rounding boundary, closer than float32 resolves, and builds disagree about which
+    # side they round on.
+    exact = [float(value) for value in torch.linalg.svdvals(MATRIX.double())]
     result = estimate_singular_values(
         MATRIX,
         n_counting_wires=N_COUNTING_WIRES,
